Add new BaseSensorTestCase class and move assert method there.
